Restructure project structure
